requirements.txt
Pillow
psutil
pystray